--- bin/main/.gitignore
@@ -1,3 +1,0 @@
-/io/
-/templates/
-/static/
