+++ bin/main/templates/buyerLogin.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8"/>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
+    <meta http-equiv="X-UA-Compatible" content="ie=edge"/>
+    <title>Document</title>
+</head>
+<body>
+    <ul>
+       <li> <input type="text" placeholder="Username" /></li>
+    <li><input type="text" placeholder="Password" /></li>
+      <li>  <button>Submit</button></li>
+    </ul>
+</body>
+</html>
