+++ bin/main/templates/farmManager.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8"/>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
+    <meta http-equiv="X-UA-Compatible" content="ie=edge"/>
+    <title>Farm Manager</title>
+</head>
+<body>
+    <h1>this is the farm manager page</h1>
+    <li> <input type="text" placeholder="Type of hay" /></li>
+    <li> <input type="text" placeholder="Type of cut" /></li>
+    <li> <input type="text" placeholder="Price per ton" /></li>
+    <li> <input type="text" placeholder="Price per bale" /></li>
+     <li> <input type="text" placeholder= "Upload Image" /></li>
+    <button>Browse</button>
+</body>
+</html>
