+++ bin/main/templates/home.html
@@ -14,8 +14,6 @@
     <h3>Changing the way you hay</h3>
     <button><a th:href="@{/farmerLogin/}">Farmer</a></button>
     <button><a th:href="@{/buyerLogin/}">Buyer</a></button>
-    <h2>Not a member? Sign up Now!</h2>
-    <button><a th:href="@{/signUpPage/}">Sign Up</a></button>
     </div>
     </div>
 </body>
