+++ bin/main/templates/index.html
@@ -29,6 +29,11 @@
             <li>
                 <a th:href="@{/index/cropList}">Crop List</a>
             </li>
+            <li>
+                    <button>
+                    <a th:href="@{/home}">Login</a>
+                    </button>
+                </li>
         </ul>
     </nav>
     <header class="header">
@@ -68,15 +73,3 @@
 </body>
 
 </html>
-© 2018 GitHub, Inc.
-Terms
-Privacy
-Security
-Status
-Help
-Contact GitHub
-Pricing
-API
-Training
-Blog
-About
