+++ src/main/java/io/retrojb/HayApp/FarmerController.java
@@ -16,7 +16,7 @@ public class FarmerController {
 	
 	@RequestMapping("/")
 	public String landingPage(Model model) {
-		return "redirect:/home";
+		return "redirect:/index";
 	}
 	
 	@RequestMapping("/home")
@@ -37,26 +37,7 @@ public class FarmerController {
 		return "farmer";
 	}
 	
-	@RequestMapping("/index/{farmerUserName}/crops/{cropId}")
-	public String getCrop(@PathVariable (name = "farmerUserName") String farmerUserName, 
-						@PathVariable (name = "cropId") Long cropId,
-									Model model) {
-		model.addAttribute("farmer", farmerRepo.findByFarmerUserName(farmerUserName));
-		model.addAttribute("crop", cropRepo.findOne(cropId));
-		return "crop";
-	}
 	
-	@RequestMapping("/index/cropList")
-	public String getAllCrops(Model model) {
-		model.addAttribute("crops", cropRepo.findAll());
-		return "cropList";
-	}
-	
-	@RequestMapping("/index/cropList/cropDetail/{cropId}")
-	public String getOneCropFromList(@PathVariable(name="cropId") Long cropId, Model model) {
-		model.addAttribute("crop", cropRepo.findOne(cropId));
-		return "cropDetail";
-	}
 	@RequestMapping("/farmerLogin")
 	public String getFarmerLogin(Model model) {
 //	model.addAttribute("farmerLogin", farmerRepo.findAll());
