+++ src/main/resources/templates/index.html
@@ -29,6 +29,11 @@
             <li>
                 <a th:href="@{/index/cropList}">Crop List</a>
             </li>
+            <li>
+                    <button>
+                    <a th:href="@{/home}">Login</a>
+                    </button>
+                </li>
         </ul>
     </nav>
     <header class="header">
